Move Category classes to Domain Category
