+++ .phpstorm.meta.php
@@ -15,8 +15,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -127,8 +127,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -239,8 +239,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -351,8 +351,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -463,8 +463,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -575,8 +575,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -687,8 +687,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -799,8 +799,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -911,8 +911,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -1023,8 +1023,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
@@ -1135,8 +1135,8 @@ namespace PHPSTORM_META {
         '' => '@',
             'App\Contracts\ArticleRepositoryInterface' => \App\Domain\Article\Repository\ArticleRepository::class,
             'App\Contracts\ArticleServiceInterface' => \App\Domain\Article\Services\ArticleService::class,
-            'App\Contracts\CategoryRepositoryInterface' => \App\Repository\CategoryRepository::class,
-            'App\Contracts\CategoryServiceInterface' => \App\Services\CategoryService::class,
+            'App\Contracts\CategoryRepositoryInterface' => \App\Domain\Category\Repository\CategoryRepository::class,
+            'App\Contracts\CategoryServiceInterface' => \App\Domain\Category\Services\CategoryService::class,
             'App\Contracts\CommentRepositoryInterface' => \App\Repository\CommentRepository::class,
             'App\Contracts\CommentServiceInterface' => \App\Services\CommentService::class,
             'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
