+++ app/Domain/Category/Category.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Domain\Category;
 
 use Database\Factories\CategoryFactory;
 use Eloquent;
