+++ app/Domain/Category/CategoryDTO.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Dto;
+namespace App\Domain\Category;
 
 class CategoryDTO
 {
