+++ app/Domain/Category/CategoryPolicy.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Policies;
+namespace App\Domain\Category;
 
 use App\Models\User;
