+++ app/Domain/Category/Requests/UpdateCategoryRequest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Requests\Category;
+namespace App\Domain\Category\Requests;
 
-use App\Dto\CategoryDTO;
-use App\Models\Category;
+use App\Domain\Category\Category;
+use App\Domain\Category\CategoryDTO;
 use Gate;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
